backend-app-api: move cors and helmet options reader to http folder
